# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Weave API
  description: Chat-based multi-bank money transfer agent API
  version: "1.0.0"
servers:
  - url: https://api.weave.io
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/transfers:
    post:
      summary: Initiate a transfer
      operationId: initiateTransfer
      tags: [Transfers]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, recipient_account, recipient_bank, recipient_name]
              properties:
                amount:
                  type: number
                  description: Amount in NGN (will be converted to kobo)
                  example: 5000
                recipient_account:
                  type: string
                  example: "0123456789"
                recipient_bank:
                  type: string
                  description: Bank code
                  example: "058"
                recipient_name:
                  type: string
                  example: "Jane Doe"
                narration:
                  type: string
                  example: "For dinner"
      responses:
        "201":
          description: Transfer initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id:
                    type: string
                    format: uuid
                  our_ref:
                    type: string
                  status:
                    type: string
                  debit_plan:
                    $ref: "#/components/schemas/DebitPlan"
        "400":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"

    get:
      summary: List transaction history
      operationId: listTransactions
      tags: [Transfers]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED]
          description: Filter by status
        - name: type
          in: query
          schema:
            type: string
            enum: [debit_leg, payout_leg, deposit, fee, refund, hold, release]
          description: Filter by transaction type
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start date (RFC3339)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End date (RFC3339)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
  /api/v1/transfers/{id}:
    get:
      summary: Get transfer by ID
      operationId: getTransfer
      tags: [Transfers]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/APIError"
  /api/v1/transfers/ref/{ref}:
    get:
      summary: Get transfer by reference
      operationId: getTransferByRef
      tags: [Transfers]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    APIError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [debit_leg, payout_leg, deposit, fee, refund, hold, release]
        amount:
          type: integer
          description: Amount in kobo
        fee:
          type: integer
          description: Fee in kobo
        currency:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED]
        source_provider:
          type: string
        recipient_account:
          type: string
        recipient_bank_code:
          type: string
        recipient_name:
          type: string
        provider_ref:
          type: string
        our_ref:
          type: string
        failure_reason:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    DebitPlan:
      type: object
      properties:
        total:
          type: integer
          description: Total amount in kobo
        fees:
          type: integer
          description: Total fees in kobo
        legs:
          type: array
          items:
            $ref: "#/components/schemas/DebitLeg"
    DebitLeg:
      type: object
      properties:
        source:
          type: string
          description: "wallet or bank account UUID"
        amount:
          type: integer
          description: Amount in kobo
        fee:
          type: integer
          description: Fee in kobo
        bank_name:
          type: string
  responses:
    APIError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"